pyproject.toml
[tool.poetry]
name = "DRF Backend Template"
version = "0.1.0"
description = "This is a template for DRF Backend projects"
authors = ["Sk Farhad <sk.farhad.eee@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
django-redis = "^5.4.0"
celery = "^5.4.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-leaflet = "^0.30.1"
djangorestframework-api-key = "^3.0.0"
django = ">=4.2.0,<4.3.0"
gevent = "^24.2.1"
django-filter = "^24.3"
firebase-admin = "^6.5.0"
django-cors-headers = "^4.4.0"
django-background-tasks = "^1.2.8"
django-rest-swagger = "^2.2.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.28.0"
drf-spectacular-sidecar = "^2025.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
